fix(advanced search): fixed advanced search, deleted unused testing files, and added testing for advacned search & chapter listing
