+++ src/app/layout/navbar/navbar.component.ts
@@ -17,7 +17,7 @@ import {
 import {faSearch} from '@fortawesome/free-solid-svg-icons';
 import {
 	ChapterListingService
-} from '../../service/chapter-listing.service';
+} from '../../service/chapter-listing/chapter-listing.service';
 import {LayoutService} from '../../service/layout.service';
 import {Router} from '@angular/router';
 import {
