--- src/app/service/layout.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { LayoutService } from './layout.service';
-
-describe('LayoutService', () => {
-  let service: LayoutService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(LayoutService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
