--- src/app/service/routing.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { RoutingService } from './routing.service';
-
-describe('ChapterRoutingService', () => {
-  let service: RoutingService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(RoutingService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
