--- src/app/service/seo.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { SeoService } from './seo.service';
-
-describe('SeoService', () => {
-  let service: SeoService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(SeoService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
