--- src/app/shared/autofocus.directive.spec.ts
@@ -1,8 +1,0 @@
-import { AutofocusDirective } from './autofocus.directive';
-
-describe('AutofocusDirective', () => {
-  it('should create an instance', () => {
-    const directive = new AutofocusDirective();
-    expect(directive).toBeTruthy();
-  });
-});
