--- src/app/shared/scroll-tracker.directive.spec.ts
@@ -1,8 +1,0 @@
-import { ScrollTrackerDirective } from './scroll-tracker.directive';
-
-describe('ScrollTrackerDirective', () => {
-  it('should create an instance', () => {
-    const directive = new ScrollTrackerDirective();
-    expect(directive).toBeTruthy();
-  });
-});
