--- src/app/shared/without-hash.pipe.spec.ts
@@ -1,8 +1,0 @@
-import { WithoutHashPipe } from './without-hash.pipe';
-
-describe('WithoutHashPipe', () => {
-  it('create an instance', () => {
-    const pipe = new WithoutHashPipe();
-    expect(pipe).toBeTruthy();
-  });
-});
